# workflow-manager/api/workflow/swagger/OpenAPI/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Workflow manager API spec",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/workflow/v1/domain/{domain}/project/{project}/workflow/{name}": {
      "get": {
        "operationId": "WorkflowApi_GetWorkflow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowWorkflowGetResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "description": "project under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "description": "name of the workflow to be deleted",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "WorkflowApi"
        ]
      },
      "delete": {
        "operationId": "WorkflowApi_DeleteWorkflow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowWorkflowDeleteResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "description": "project under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "description": "name of the workflow to be deleted",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "WorkflowApi"
        ]
      }
    },
    "/api/workflow/v1/domain/{domain}/project/{project}/workflows": {
      "get": {
        "summary": "List Workflows executed in a project",
        "operationId": "WorkflowApi_ListWorkflows",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowWorkflowListResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "description": "project under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "offset",
            "description": "offset for pagination.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "description": "limit for pagination.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "search",
            "description": "search text.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WorkflowApi"
        ]
      }
    }
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "workflowTemplateLink": {
      "type": "object",
      "properties": {
        "source": {
          "type": "string",
          "title": "source - nodeid of one of the nodes above"
        },
        "sourceVar": {
          "type": "string",
          "title": "source variable - output from the source"
        },
        "target": {
          "type": "string",
          "title": "target - nodeid of one of the nodes above"
        },
        "targetVar": {
          "type": "string",
          "title": "target variable - input of the target"
        }
      }
    },
...
    "workflowWorkflowDefStatus": {
      "type": "string",
      "enum": [
        "Created",
        "Scheduled",
        "Running",
        "Completed",
        "Failed"
      ],
      "default": "Created",
      "title": "- Created: workflow is created by pending to be scheduled\n - Scheduled: workflow is currently scheduled for execution\n - Running: workflow is currently being executed\n - Completed: workflow execution is completed\n - Failed: workflow execution failed"
    },
    "workflowWorkflowDeleteResp": {
      "type": "object"
    },
    "workflowWorkflowGetResp": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "name of the workflow execution"
        },
        "template": {
          "type": "string",
          "title": "workflow template used"
        },
        "desc": {
          "type": "string",
          "title": "workflow descrption"
        },
        "status": {
          "$ref": "#/definitions/workflowWorkflowDefStatus",
          "title": "status of the workflow"
        },
        "createdBy": {
          "type": "string",
          "title": "executed by"
        },
        "createTime": {
          "type": "string",
          "format": "int64",
          "title": "create time"
        },
        "startTime": {
          "type": "string",
          "format": "int64",
          "title": "start time"
        },
        "endTime": {
          "type": "string",
          "format": "int64",
          "title": "completion time"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "optional tags"
        },
        "isDeleted": {
          "type": "boolean",
          "title": "is this entry being deleted"
        },
        "nodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowWorkflowNode"
          },
          "title": "workflow nodes"
        },
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowTemplateLink"
          },
          "title": "workflow links"
        }
      }
    },
    "workflowWorkflowListEntry": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "name of the workflow execution"
        },
        "template": {
          "type": "string",
          "title": "workflow template used"
        },
        "desc": {
          "type": "string",
          "title": "workflow descrption"
        },
        "status": {
          "$ref": "#/definitions/workflowWorkflowDefStatus",
          "title": "status of the workflow"
        },
        "createdBy": {
          "type": "string",
          "title": "executed by"
        },
        "createTime": {
          "type": "string",
          "format": "int64",
          "title": "create time"
        },
        "startTime": {
          "type": "string",
          "format": "int64",
          "title": "start time"
        },
        "endTime": {
          "type": "string",
          "format": "int64",
          "title": "completion time"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "optional tags"
        },
        "isDeleted": {
          "type": "boolean",
          "title": "is this entry being deleted"
        }
      }
    },
    "workflowWorkflowListResp": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "count of total workflows executed in this project"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowWorkflowListEntry"
          },
          "title": "list of items for current response"
        }
      }
    },
    "workflowWorkflowNode": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/workflowWorkflowNodeNodeType",
          "title": "type of the node"
        },
        "nodeId": {
          "type": "string",
          "title": "node unique id in the template\nto identify links"
        },
        "moduleId": {
          "type": "string",
          "title": "unique id of the module or catalog entry"
        },
        "name": {
          "type": "string",
          "title": "name of the node"
        },
        "x": {
          "type": "number",
          "format": "double",
          "title": "position x"
        },
        "y": {
          "type": "number",
          "format": "double",
          "title": "position y"
        },
        "status": {
          "$ref": "#/definitions/workflowWorkflowDefStatus",
          "title": "status of node"
        },
        "err": {
          "type": "string",
          "title": "error message"
        },
        "inputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowWorkflowNodeData"
          },
          "title": "inputs"
        },
        "outputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowWorkflowNodeData"
          },
          "title": "outputs"
        }
      }
    },
    "workflowWorkflowNodeData": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "name of the data"
        },
        "value": {
          "type": "string",
          "title": "value for the data"
        }
      }
    },
    "workflowWorkflowNodeNodeType": {
      "type": "string",
      "enum": [
        "Module",
        "Catalog",
        "UserInput"
      ],
      "default": "Module",
      "description": "- Module: module node, where domain and project is assumed\nsame as template\n - Catalog: module from catalog, where the domain is assumed\nsame as template, and project is considered empty\n - UserInput: user input, where user will be prompted for input\nat the time of executing the workflow template",
      "title": "node type"
    }
  },
...
}